% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K Nearest Neighbor Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Dataframe with training data, without response variable.}

\item{cl}{Factor of response variables.}

\item{k_nn}{Integer representing number of nearest neighbors.}

\item{k_cv}{Integer for number of folds in cross validation.}
}
\value{
List with elements:
class - Factor of classification using the training data,
cv_err - Numeric representing cross-validation error.
}
\description{
Performs a KNN cross-validation test.
}
\examples{
my_knn_cv(iris[1:4], iris$Species, 5, 5)
my_knn_cv(iris[1:4], iris$Species, 1, 5)

}
\keyword{prediction}
