% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Fits a Linear Model}
\usage{
my_lm(lm_form, lm_data)
}
\arguments{
\item{lm_form}{Formula used to fit model.}

\item{lm_data}{Numeric data used to fit model.}
}
\value{
List with values:
\code{Estimate} - Numeric coefficient of gitted model,
\code{Std. Error} - Numeric standard error of coefficient,
\code{t value} - Numeric test-statistic of coefficient,
\code{Pr(>|t|)} - Numeric p-value of coefficient.
}
\description{
This function fits a linear model using the data.
}
\examples{
my_lm(speed ~ dist, cars)

}
\keyword{inference}
\keyword{prediction}
