% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric input representing number of folds for cross-validation.}
}
\value{
Numeric representing MSE of the random forest.
}
\description{
Performs a random forest cross-validation test specific to the
\code{palmerpenguins} dataset, with variables
\code{bill_depth_mm}, \code{bill_length_mm}, \code{flipper_length_mm},
with response variable \code{body_mass_g}.
}
\examples{
my_rf_cv(5)

}
\keyword{prediction}
