% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Fits a Linear Model}
\usage{
my_lm(lm_form, lm_data)
}
\arguments{
\item{lm_form}{Formula used to fit model.}

\item{lm_data}{Numeric data used to fit model.}
}
\value{
List with values:
Estimate - numeric coefficient of line of best fit,
Std. Error - numeric representing the standard error of coefficients,
t value - numeric test-statistic of coefficient,
Pr(>|t|) - numeric p-value of coefficient.
}
\description{
This function fits a linear model using the data.
}
\examples{
my_lm(speed ~ dist, cars)

}
