% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric input representing number of folds for cross-validation.}
}
\value{
Numeric representing average MSE of the k-fold simulations.
}
\description{
Performs a random forest cross-validation test specific to the
\code{my_penguins} dataset, with variables
\code{bill_depth_mm}, \code{bill_length_mm}, \code{flipper_length_mm}
and response variable \code{body_mass_g}.
}
\examples{
my_rf_cv(5)
my_rf_cv(3)

}
\keyword{prediction}
