% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{t-test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric input of data.}

\item{alternative}{Character input, either \code{"greater"},
\code{"less"}, or \code{"two.sided"}, corresponding to what hypothesis
is being tested.}

\item{mu}{The numeric null hypothesis value.}
}
\value{
List with values:
test_stat - numeric representing test_statistic,
df - numeric representing degrees of freedom,
alternative - character either "greater", "less", or "two.sided", representing
type of t-test,
p_val - numeric representing the p-value.
}
\description{
This function performs a one sample t-test.
}
\examples{
my_t.test(c(1:5), "greater", 4)
my_t.test(c(1:10), "less", 4)
my_t.test(c(1:5), "two.sided", 0)

}
